split ipfs fetch functions to utils
