+++ src/lib/utils/create-fetch-ipfs.ts
@@ -1,7 +1,3 @@
 const IPFS = require('ipfs');
 
-export const createFetchIpfs = async () => {
-  const node = await IPFS.create();
-
-  return node;
-};
+export const createFetchIpfs = async () => await IPFS.create();
