+++ src/lib/utils/index.ts
@@ -3,3 +3,5 @@ export * from './fetch-rmrk-metadata';
 export * from './sanitize-ipfs-url';
 export * from './get-ipfs-cid';
 export * from './resolve-ipfs-provider';
+export * from './get-ipfs-json';
+export * from './get-ipfs-image';
